requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
blessed==1.20.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
cryptography==43.0.1
discord.py==2.4.0
distro==1.9.0
dj-rest-auth==6.0.0
Django==4.2
django-allauth==65.0.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-picklefield==3.2
django-seed==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==30.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
msgpack==1.1.0
multidict==6.1.0
oauthlib==3.2.2
openai==0.28.0
pillow==10.4.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.6.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
toposort==1.10
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.14.0
